+++ holehe/modules/shopping/deliveroo.py
@@ -1,0 +1,34 @@
+import requests
+import random
+from bs4 import BeautifulSoup
+
+from holehe.localuseragent import ua
+
+
+def deliveroo(email):
+
+    headers = {
+        'User-Agent': random.choice(ua["browsers"]["chrome"]),
+        'Accept': 'application/json, application/vnd.api+json',
+        'Accept-Language': 'en-US;q=0.5,en;q=0.3',
+        'X-Requested-With': 'XMLHttpRequest',
+        'X-Roo-Client': 'orderweb-client',
+        'X-Roo-Country': 'fr',
+        'Content-Type': 'application/json;charset=UTF-8',
+        'Origin': 'https://deliveroo.com',
+        'DNT': '1',
+        'Connection': 'keep-alive',
+        'TE': 'Trailers',
+    }
+
+    data = '{"email_address":"tazeest@gmail.com"}'
+
+    response = requests.post('https://consumer-ow-api.deliveroo.com/orderapp/v1/check-email', headers=headers, data=data)
+    if response.status_code==200:
+        data=response.json()
+        if data["registered"]==True:
+            return({"rateLimit": False, "exists": True, "emailrecovery": None, "phoneNumber": None, "others": None})
+        else:
+            return({"rateLimit": False, "exists": False, "emailrecovery": None, "phoneNumber": None, "others": None})
+    else:
+            return({"rateLimit": True, "exists": False, "emailrecovery": None, "phoneNumber": None, "others": None})
