+++ holehe/modules/social_media/facebook.py
@@ -1,6 +1,7 @@
 import requests
 import random
 import json
+import re
 
 from bs4 import BeautifulSoup
 from holehe.localuseragent import ua
