+++ holehe/modules/social_media/plurk.py
@@ -1,0 +1,27 @@
+import requests
+import random
+from holehe.localuseragent import ua
+
+def plurk(email):
+    headers = {
+        'User-Agent': random.choice(ua["browsers"]["firefox"]),
+        'Accept': '*/*',
+        'Accept-Language': 'fr,fr-FR;q=0.8,en-US;q=0.5,en;q=0.3',
+        'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8',
+        'X-Requested-With': 'XMLHttpRequest',
+        'Origin': 'https://www.plurk.com',
+        'DNT': '1',
+        'Connection': 'keep-alive',
+    }
+
+    data = {
+      'email': email
+    }
+
+    response = requests.post('https://www.plurk.com/Users/isEmailFound', headers=headers, data=data)
+    if response.text=="True":
+        return({"rateLimit": False, "exists": True, "emailrecovery": None, "phoneNumber": None, "others": None})
+    elif response.text=="False":
+        return({"rateLimit": False, "exists": False, "emailrecovery": None, "phoneNumber": None, "others": None})
+    else:
+        return({"rateLimit": True, "exists": False, "emailrecovery": None, "phoneNumber": None, "others": None})
